requirements.txt
torch>=2.0
torchvision>=0.15
numpy
pyyaml
Pillow
scikit-learn
opencv-python
submitit
